refactor(rect): move iters to module and add docs
